refactor: Move createCustomReportCollection() each analyzer to custom_report_collection
