+++ src/analyzer/github_analyzer.ts
@@ -3,8 +3,7 @@ import { sumBy, min, max } from 'lodash'
 import { Analyzer, diffSec, Status, TestReport, WorkflowParams, convertToReportTestSuites } from './analyzer'
 import { RepositoryTagMap } from '../client/github_repository_client'
 import { parse } from 'junit2json'
-import { CustomReportArtifact, Artifact } from '../client/client'
-import { CustomReportCollection, CustomReport } from '../custom_report_collection'
+import { Artifact } from '../client/client'
 export type WorkflowRunsItem = RestEndpointMethodTypes['actions']['listWorkflowRunsForRepo']['response']['data']['workflow_runs'][0]
 export type JobsItem = RestEndpointMethodTypes['actions']['listJobsForWorkflowRun']['response']['data']['jobs']
 
@@ -178,31 +177,5 @@ export class GithubAnalyzer implements Analyzer {
       }
     }
     return testReports
-  }
-
-  async createCustomReportCollection(workflowReport: WorkflowReport, customReportArtifacts: CustomReportArtifact): Promise<CustomReportCollection> {
-    const reportCollection = new CustomReportCollection()
-    for (const [reportName, artifacts] of customReportArtifacts) {
-      const reports = artifacts.map((artifact) => {
-        let data: { [key: string]: unknown }
-        try {
-          const jsonString = Buffer.from(artifact.data).toString('utf8')
-          data = JSON.parse(jsonString)
-        } catch (error) {
-          console.error(`Error: Could not parse as JSON. ${artifact.path}`)
-          return
-        }
-
-        return {
-          workflowId: workflowReport.workflowId,
-          workflowRunId: workflowReport.workflowRunId,
-          createdAt: workflowReport.createdAt,
-          ...data
-        } as CustomReport
-      }).filter((report) => report !== undefined) as CustomReport[]
-
-      reportCollection.set(reportName, reports)
-    }
-    return reportCollection
   }
 }
