+++ src/custom_report_collection.ts
@@ -1,8 +1,37 @@
+import { WorkflowReport } from "./analyzer/analyzer"
+import { CustomReportArtifact } from "./client/client"
+
 export type CustomReport = {
   workflowId: string
   workflowRunId: string
   createdAt: Date,
   [key: string]: unknown
+}
+
+export const createCustomReportCollection = async (workflowReport: WorkflowReport, customReportArtifacts: CustomReportArtifact): Promise<CustomReportCollection> => {
+  const reportCollection = new CustomReportCollection()
+  for (const [reportName, artifacts] of customReportArtifacts) {
+    const reports = artifacts.map((artifact) => {
+      let data: { [key: string]: unknown }
+      try {
+        const jsonString = Buffer.from(artifact.data).toString('utf8')
+        data = JSON.parse(jsonString)
+      } catch (error) {
+        console.error(`Error: Could not parse as JSON. ${artifact.path}`)
+        return
+      }
+
+      return {
+        workflowId: workflowReport.workflowId,
+        workflowRunId: workflowReport.workflowRunId,
+        createdAt: workflowReport.createdAt,
+        ...data
+      } as CustomReport
+    }).filter((report) => report !== undefined) as CustomReport[]
+
+    reportCollection.set(reportName, reports)
+  }
+  return reportCollection
 }
 
 export class CustomReportCollection {
